[Fix] API用のエラーハンドリングをconcernに切り出し
