+++ app/controllers/line/base_controller.rb
@@ -3,6 +3,7 @@
 module Line
   class BaseController < ApplicationController
     require 'line/bot'
+    include Api::ExceptionHandler
 
     before_action :validate_signature
